refactor: 使用枚举类重构devicon导入
